Move useAbortController into the framework as it is needed by framework components
