+++ frontend/awx/views/schedules/hooks/useProcessCredentials.tsx
@@ -1,11 +1,11 @@
 import { useCallback } from 'react';
-import { useAbortController } from '../../../../common/crud/useAbortController';
+import { useAbortController } from '../../../../../framework/hooks/useAbortController';
 import { usePostRequest } from '../../../../common/crud/usePostRequest';
 import { awxAPI } from '../../../common/api/awx-utils';
-import { PromptFormValues } from '../../../resources/templates/WorkflowVisualizer/types';
-import { Credential } from '../../../interfaces/Credential';
 import { getAddedAndRemoved } from '../../../common/util/getAddedAndRemoved';
+import { Credential } from '../../../interfaces/Credential';
 import { LaunchConfiguration } from '../../../interfaces/LaunchConfiguration';
+import { PromptFormValues } from '../../../resources/templates/WorkflowVisualizer/types';
 
 export const useProcessCredentials = () => {
   const abortController = useAbortController();
