+++ frontend/awx/views/schedules/hooks/useProcessInstanceGroups.tsx
@@ -1,11 +1,11 @@
 import { useCallback } from 'react';
-import { useAbortController } from '../../../../common/crud/useAbortController';
+import { useAbortController } from '../../../../../framework/hooks/useAbortController';
 import { usePostRequest } from '../../../../common/crud/usePostRequest';
-import { getAddedAndRemoved } from '../../../common/util/getAddedAndRemoved';
 import { awxAPI } from '../../../common/api/awx-utils';
-import { PromptFormValues } from '../../../resources/templates/WorkflowVisualizer/types';
+import { getAddedAndRemoved } from '../../../common/util/getAddedAndRemoved';
 import { InstanceGroup } from '../../../interfaces/InstanceGroup';
 import { LaunchConfiguration } from '../../../interfaces/LaunchConfiguration';
+import { PromptFormValues } from '../../../resources/templates/WorkflowVisualizer/types';
 
 export const useProcessInstanceGroups = () => {
   const abortController = useAbortController();
